Move http exceptions into the http package.
This will assist in making the Http package a subsplit, and also puts
the exceptions in the same package as where they are frequently used.
