+++ src/Http/Exception/InternalErrorException.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ * @since         3.0.0
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Cake\Http\Exception;
+
+/**
+ * Represents an HTTP 500 error.
+ */
+class InternalErrorException extends HttpException
+{
+
+    /**
+     * Constructor
+     *
+     * @param string|null $message If no message is given 'Internal Server Error' will be the message
+     * @param int $code Status code, defaults to 500
+     * @param \Exception|null $previous The previous exception.
+     */
+    public function __construct($message = null, $code = null, $previous = null)
+    {
+        if (empty($message)) {
+            $message = 'Internal Server Error';
+        }
+        parent::__construct($message, $code, $previous);
+    }
+}
