+++ src/Http/Exception/NotImplementedException.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ * @since         3.0.0
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Cake\Http\Exception;
+
+/**
+ * Not Implemented Exception - used when an API method is not implemented
+ */
+class NotImplementedException extends HttpException
+{
+
+    /**
+     * {@inheritDoc}
+     */
+    protected $_messageTemplate = '%s is not implemented.';
+
+    /**
+     * {@inheritDoc}
+     */
+    protected $_defaultCode = 501;
+}
