+++ src/Network/Exception/ForbiddenException.php
@@ -1,40 +1,4 @@
 <?php
-/**
- * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- *
- * Licensed under The MIT License
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- * @since         3.0.0
- * @license       https://opensource.org/licenses/mit-license.php MIT License
- */
-namespace Cake\Network\Exception;
-
-/**
- * Represents an HTTP 403 error.
- */
-class ForbiddenException extends HttpException
-{
-
-    /**
-     * {@inheritDoc}
-     */
-    protected $_defaultCode = 403;
-
-    /**
-     * Constructor
-     *
-     * @param string|null $message If no message is given 'Forbidden' will be the message
-     * @param int $code Status code, defaults to 403
-     * @param \Exception|null $previous The previous exception.
-     */
-    public function __construct($message = null, $code = null, $previous = null)
-    {
-        if (empty($message)) {
-            $message = 'Forbidden';
-        }
-        parent::__construct($message, $code, $previous);
-    }
-}
+// @deprecated Backward compatibility alias
+class_alias('Cake\Http\Exception\ForbiddenException', 'Cake\Network\Exception\ForbiddenException');
+deprecationWarning('Use Cake\Http\Exception\ForbiddenException instead of Cake\Network\Exception\ForbiddenException.');
