+++ src/Network/Exception/HttpException.php
@@ -1,27 +1,4 @@
 <?php
-/**
- * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- *
- * Licensed under The MIT License
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- * @since         3.0.0
- * @license       https://opensource.org/licenses/mit-license.php MIT License
- */
-namespace Cake\Network\Exception;
-
-use Cake\Core\Exception\Exception;
-
-/**
- * Parent class for all of the HTTP related exceptions in CakePHP.
- * All HTTP status/error related exceptions should extend this class so
- * catch blocks can be specifically typed.
- *
- * You may also use this as a meaningful bridge to Cake\Core\Exception\Exception, e.g.:
- * throw new \Cake\Network\Exception\HttpException('HTTP Version Not Supported', 505);
- */
-class HttpException extends Exception
-{
-}
+// @deprecated Backward compatibility alias
+class_alias('Cake\Http\Exception\HttpException', 'Cake\Network\Exception\HttpException');
+deprecationWarning('Use Cake\Http\Exception\HttpException instead of Cake\Network\Exception\HttpException.');
