+++ src/Network/Exception/ServiceUnavailableException.php
@@ -1,40 +1,4 @@
 <?php
-/**
- * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- *
- * Licensed under The MIT License
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
- * @since         3.1.7
- * @license       https://opensource.org/licenses/mit-license.php MIT License
- */
-namespace Cake\Network\Exception;
-
-/**
- * Represents an HTTP 503 error.
- */
-class ServiceUnavailableException extends HttpException
-{
-
-    /**
-     * {@inheritDoc}
-     */
-    protected $_defaultCode = 503;
-
-    /**
-     * Constructor
-     *
-     * @param string|null $message If no message is given 'Service Unavailable' will be the message
-     * @param int $code Status code, defaults to 503
-     * @param \Exception|null $previous The previous exception.
-     */
-    public function __construct($message = null, $code = null, $previous = null)
-    {
-        if (empty($message)) {
-            $message = 'Service Unavailable';
-        }
-        parent::__construct($message, $code, $previous);
-    }
-}
+// @deprecated Backward compatibility alias
+class_alias('Cake\Http\Exception\ServiceUnavailableException', 'Cake\Network\Exception\ServiceUnavailableException');
+deprecationWarning('Use Cake\Http\Exception\ServiceUnavailableException instead of Cake\Network\Exception\ServiceUnavailableException.');
